+++ app/includes/lib/Security.php
@@ -2,21 +2,18 @@
 
 class Security {
 
-	public function __construct() {
+	protected static $token_name = 'csrf_token';
 
-		if (empty(Session::get('token'))) {
+	public function __construct() {
+		if (empty(Session::get(self::$token_name))) {
 			self::generateCSRFToken();
 		}
-
 		self::verifyToken();
 	}
 
 	public static function verifyToken() {
-
-		$token_name = 'token';
-
 		if ($_SERVER['REQUEST_METHOD'] !== 'GET') {
-			if (!isset($_REQUEST[$token_name]) || $_REQUEST[$token_name] !== Session::get('token')) {
+			if (!isset($_REQUEST[self::$token_name]) || $_REQUEST[self::$token_name] !== Session::get(self::$token_name)) {
 				header('HTTP/1.1 403 Forbidden');
 				exit();
 			}
@@ -25,10 +22,14 @@ class Security {
 
 	public static function generateCSRFToken() {
 		if (function_exists('mcrypt_create_iv')) {
-			Session::set('token', bin2hex(mcrypt_create_iv(32, MCRYPT_DEV_URANDOM)));
+			Session::set(self::$token_name, bin2hex(mcrypt_create_iv(32, MCRYPT_DEV_URANDOM)));
 		} else {
-			Session::set('token', bin2hex(openssl_random_pseudo_bytes(32)));
+			Session::set(self::$token_name, bin2hex(openssl_random_pseudo_bytes(32)));
 		}
+	}
+
+	public static function renderCSRFInput() {
+		return '<input type="hidden" name="' . Security::$token_name .'" value="' . Session::get(Security::$token_name) . '">';
 	}
 
 }